cleanup

+++ 15Puzzle/src/main/java/de/amr/games/puzzle15/PuzzleSimulator.java
@@ -14,7 +14,7 @@ public class PuzzleSimulator extends JFrame {
 	private PuzzleView view;
 
 	public PuzzleSimulator() {
-		puzzle = new Puzzle((byte) 4);
+		puzzle = new Puzzle(4);
 		view = new PuzzleView(puzzle, 100);
 		setTitle("15-Puzzle");
 		setDefaultCloseOperation(EXIT_ON_CLOSE);
