+++ 15Puzzle/src/test/java/de/amr/games/puzzle15/test/PuzzleTest.java
@@ -14,7 +14,7 @@ public class PuzzleTest {
 
 	@Before
 	public void setUp() {
-		puzzle = new Puzzle((byte) 4);
+		puzzle = new Puzzle(4);
 	}
 
 	@Test
